requirements.txt
Flask
gunicorn
Pillow
google-generativeai
python-dotenv
google-cloud-vision